feat(profile-addons-api): fixed comments

+++ src/app/shared/services/addons/addons.service.ts
@@ -1,4 +1,4 @@
-import { Store } from '@ngxs/store';
+import { select } from '@ngxs/store';
 
 import { map, Observable } from 'rxjs';
 
@@ -27,18 +27,17 @@ import {
 } from '@shared/models';
 import { OperationInvocationResponseJsonApi } from '@shared/models/addons/operation-invocation.models';
 
-import { environment } from '../../../../environments/environment';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AddonsService {
-  #store = inject(Store);
-  #jsonApiService = inject(JsonApiService);
-  #currentUser = this.#store.selectSignal(UserSelectors.getCurrentUser);
+  private jsonApiService = inject(JsonApiService);
+  private currentUser = select(UserSelectors.getCurrentUser);
 
   getAddons(addonType: string): Observable<Addon[]> {
-    return this.#jsonApiService
+    return this.jsonApiService
       .get<
         JsonApiResponse<AddonGetResponseJsonApi[], null>
       >(`${environment.addonsApiUrl}/external-${addonType}-services`)
@@ -50,26 +49,26 @@ export class AddonsService {
   }
 
   getAddonsUserReference(): Observable<UserReferenceJsonApi[]> {
-    const currentUser = this.#currentUser();
+    const currentUser = this.currentUser();
     if (!currentUser) throw new Error('Current user not found');
 
-    const userUri = `https://staging4.osf.io/${currentUser.id}`;
+    const userUri = `${environment.baseResourceUri}${currentUser.id}`;
     const params = {
       'filter[user_uri]': userUri,
     };
 
-    return this.#jsonApiService
+    return this.jsonApiService
       .get<JsonApiResponse<UserReferenceJsonApi[], null>>(environment.addonsApiUrl + '/user-references/', params)
       .pipe(map((response) => response.data));
   }
 
   getAddonsResourceReference(resourceId: string): Observable<ResourceReferenceJsonApi[]> {
-    const resourceUri = `https://staging4.osf.io/${resourceId}`;
+    const resourceUri = `${environment.baseResourceUri}${resourceId}`;
     const params = {
       'filter[resource_uri]': resourceUri,
     };
 
-    return this.#jsonApiService
+    return this.jsonApiService
       .get<
         JsonApiResponse<ResourceReferenceJsonApi[], null>
       >(environment.addonsApiUrl + '/resource-references/', params)
@@ -80,7 +79,7 @@ export class AddonsService {
     const params = {
       [`fields[external-${addonType}-services]`]: 'external_service_name',
     };
-    return this.#jsonApiService
+    return this.jsonApiService
       .get<
         JsonApiResponse<AuthorizedAddonGetResponseJsonApi[], IncludedAddonData[]>
       >(`${environment.addonsApiUrl}/user-references/${referenceId}/authorized_${addonType}_accounts/?include=external-${addonType}-service`, params)
@@ -92,7 +91,7 @@ export class AddonsService {
   }
 
   getConfiguredAddons(addonType: string, referenceId: string): Observable<ConfiguredAddon[]> {
-    return this.#jsonApiService
+    return this.jsonApiService
       .get<
         JsonApiResponse<ConfiguredAddonGetResponseJsonApi[], null>
       >(`${environment.addonsApiUrl}/resource-references/${referenceId}/configured_${addonType}_addons/`)
@@ -107,7 +106,7 @@ export class AddonsService {
     addonRequestPayload: AuthorizedAddonRequestJsonApi,
     addonType: string
   ): Observable<AuthorizedAddonResponseJsonApi> {
-    return this.#jsonApiService.post<AuthorizedAddonResponseJsonApi>(
+    return this.jsonApiService.post<AuthorizedAddonResponseJsonApi>(
       `${environment.addonsApiUrl}/authorized-${addonType}-accounts/`,
       addonRequestPayload
     );
@@ -118,7 +117,7 @@ export class AddonsService {
     addonType: string,
     addonId: string
   ): Observable<AuthorizedAddonResponseJsonApi> {
-    return this.#jsonApiService.patch<AuthorizedAddonResponseJsonApi>(
+    return this.jsonApiService.patch<AuthorizedAddonResponseJsonApi>(
       `${environment.addonsApiUrl}/authorized-${addonType}-accounts/${addonId}/`,
       addonRequestPayload
     );
@@ -128,7 +127,7 @@ export class AddonsService {
     addonRequestPayload: ConfiguredAddonRequestJsonApi,
     addonType: string
   ): Observable<ConfiguredAddonResponseJsonApi> {
-    return this.#jsonApiService.post<ConfiguredAddonResponseJsonApi>(
+    return this.jsonApiService.post<ConfiguredAddonResponseJsonApi>(
       `${environment.addonsApiUrl}/configured-${addonType}-addons/`,
       addonRequestPayload
     );
@@ -139,7 +138,7 @@ export class AddonsService {
     addonType: string,
     addonId: string
   ): Observable<ConfiguredAddonResponseJsonApi> {
-    return this.#jsonApiService.patch<ConfiguredAddonResponseJsonApi>(
+    return this.jsonApiService.patch<ConfiguredAddonResponseJsonApi>(
       `${environment.addonsApiUrl}/configured-${addonType}-addons/${addonId}/`,
       addonRequestPayload
     );
@@ -148,7 +147,7 @@ export class AddonsService {
   createAddonOperationInvocation(
     invocationRequestPayload: OperationInvocationRequestJsonApi
   ): Observable<OperationInvocation> {
-    return this.#jsonApiService
+    return this.jsonApiService
       .post<OperationInvocationResponseJsonApi>(
         `${environment.addonsApiUrl}/addon-operation-invocations/`,
         invocationRequestPayload
@@ -161,10 +160,10 @@ export class AddonsService {
   }
 
   deleteAuthorizedAddon(id: string, addonType: string): Observable<void> {
-    return this.#jsonApiService.delete(`${environment.addonsApiUrl}/authorized-${addonType}-accounts/${id}/`);
+    return this.jsonApiService.delete(`${environment.addonsApiUrl}/authorized-${addonType}-accounts/${id}/`);
   }
 
   deleteConfiguredAddon(id: string, addonType: string): Observable<void> {
-    return this.#jsonApiService.delete(`${environment.addonsApiUrl}/${addonType}/${id}/`);
+    return this.jsonApiService.delete(`${environment.addonsApiUrl}/${addonType}/${id}/`);
   }
 }
